# specs/001-coil-counter-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ZootBox Coil-Counter Backend API
  description: |
    RESTful API for ZootBox vending machine inventory management and vend operations.

    **Integration**: This API is consumed by the existing Android app (MyApplication/) running on the same Android tablet via localhost.

    **Authentication**: None required (localhost-only API, physical access is security boundary).

    **Base URL**: http://127.0.0.1:8080

    **Correlation IDs**: All requests should include `X-Correlation-ID` header for distributed tracing. Backend will generate one if not provided.
  version: 1.0.0
  contact:
    name: ZootBox Development Team
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local Android tablet (localhost only)
...
paths:
...
  /api/v1/coils/{coilId}:
    get:
      tags: [Inventory]
      summary: Get single coil inventory
      description: Returns inventory state for a specific coil.
      operationId: getCoil
      parameters:
        - $ref: '#/components/parameters/CoilID'
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '200':
          description: Coil inventory details
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coil'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/coils/low-stock:
    get:
      tags: [Inventory]
      summary: Get low-stock coils for monitoring dashboard
      description: Returns coils with inventory ≤ 2 for operator dashboard and alerts
      operationId: getLowStockCoils
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '200':
          description: Low-stock coils
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  coils:
                    type: array
                    items:
                      $ref: '#/components/schemas/Coil'
                    description: Coils with inventory ≤ 2
                  low_stock_threshold:
                    type: integer
                    example: 2
                    description: Threshold value for low stock
                  count:
                    type: integer
                    example: 5
                    description: Number of low-stock coils
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    CoilID:
      name: coilId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-J](1[0]|[1-9])$"
      description: Coil ID (e.g., A5, J10)
      example: "A5"
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Request correlation ID for distributed tracing
      example: "550e8400-e29b-41d4-a716-446655440000"
  headers:
    CorrelationID:
      description: Response correlation ID (echoed from request or generated)
      schema:
        type: string
        format: uuid
  schemas:
    Coil:
      type: object
      required: [id, inventory, status, version, updated_at]
      properties:
        id:
          type: string
          example: "A5"
          pattern: "^[A-J](1[0]|[1-9])$"
        inventory:
          type: integer
          minimum: 0
          maximum: 10
          example: 7
        status:
          type: string
          enum: [available, jammed]
          example: "available"
        version:
          type: integer
          example: 42
          description: Optimistic locking version (incremented on every update)
        link_group_id:
          type: string
          format: uuid
          nullable: true
          example: null
        updated_at:
          type: string
          format: date-time
          example: "2025-12-27T10:30:00Z"
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "invalid_request"
        message:
          type: string
          example: "Coil ID must match pattern A-J followed by 1-10"
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'